refactor(core): move BlockArray to tenax.core, defer Cython import
Item 3 of #351. The numpy-backed ``BlockArray`` lived under
``tenax.algorithms._block_array`` but is imported by ``tenax.linalg``
inside ``svd``/``qr``, and ``BlockArray`` itself imported the Cython
BLAS helpers from ``tenax.contraction._cython_blas`` at module load.
That formed an ``algorithms → linalg → contraction → algorithms``
round trip in the static import graph.

Two changes break the cycle:

1. Move the module to ``tenax.core._block_array``.  The class only
   uses ``tenax.core.index.TensorIndex`` at module level, so this is
   the natural lowest-layer home and lets both ``linalg`` and the
   algorithm layer depend on it without going through ``algorithms``.

2. Defer the optional ``tenax.contraction._cython_blas`` import to
   first use via ``_ensure_cython_ba_loaded()``.  The static AST graph
   of ``_block_array`` no longer has any edge into ``contraction``, so
   the cross-layer cycle disappears even though the runtime fast path
   (BLAS ddot / daxpy / dscal) is unchanged.

All in-tree consumers were updated to the new import path; no public
``tenax.algorithms.*`` API was affected (the module had a leading
underscore).
